fix(interface-manager): move netlink logic out of net
This never truly belonged here.
