+++ net/src/interface/mod.rs
@@ -8,26 +8,18 @@
 
 //! Data structures and methods for interacting with / describing network interfaces
 
-use crate::eth::ethtype::EthType;
 use crate::eth::mac::SourceMac;
-use crate::route::RouteTableId;
-use crate::vxlan::InvalidVni;
 use derive_builder::Builder;
 use multi_index_map::MultiIndexMap;
-use rtnetlink::packet_route::link::{
-    InfoBridge, InfoData, InfoVrf, InfoVxlan, LinkAttribute, LinkFlags, LinkInfo, LinkMessage,
-    State,
-};
 use serde::{Deserialize, Serialize};
 use std::convert::TryFrom;
 use std::fmt::{Debug, Display, Formatter};
-use tracing::{error, warn};
+use tracing::error;
 
 mod bridge;
 mod vrf;
 mod vtep;
 
-use crate::ipv4::UnicastIpv4Addr;
 #[allow(unused_imports)] // re-export
 pub use bridge::*;
 #[allow(unused_imports)] // re-export
@@ -263,169 +255,6 @@ pub enum InterfaceProperties {
     Vrf(VrfProperties),
     /// Properties of something we don't currently support manipulating
     Other,
-}
-fn extract_vrf_data(builder: &mut VrfPropertiesBuilder, info: &LinkInfo) {
-    if let LinkInfo::Data(InfoData::Vrf(datas)) = info {
-        for data in datas {
-            if let InfoVrf::TableId(raw) = data {
-                match RouteTableId::try_from(*raw) {
-                    Ok(route_table) => {
-                        builder.route_table_id(route_table);
-                    }
-                    Err(err) => {
-                        error!("zero is not a legal route table id!: {err:?}");
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn extract_vxlan_info(builder: &mut VtepPropertiesBuilder, info: &LinkInfo) {
-    if let LinkInfo::Data(InfoData::Vxlan(datas)) = info {
-        for data in datas {
-            match data {
-                InfoVxlan::Id(vni) => {
-                    match (*vni).try_into() {
-                        Ok(vni) => {
-                            builder.vni(Some(vni));
-                        }
-                        Err(InvalidVni::ReservedZero) => {
-                            builder.vni(None); // likely an external vtep
-                        }
-                        Err(InvalidVni::TooLarge(wrong)) => {
-                            error!("found too large VNI: {wrong}");
-                        }
-                    }
-                }
-                InfoVxlan::Local(local) => {
-                    if local.is_unspecified() {
-                        warn!("likely os error: vtep local is 0.0.0.0.  Ignoring");
-                        builder.local(None);
-                    } else {
-                        match UnicastIpv4Addr::try_from(*local) {
-                            Ok(local) => {
-                                builder.local(Some(local));
-                            }
-                            Err(err) => {
-                                error!("{err}");
-                                builder.local(None);
-                            }
-                        }
-                    }
-                }
-                InfoVxlan::Ttl(ttl) => {
-                    builder.ttl(Some(*ttl));
-                }
-                _ => {}
-            }
-        }
-    }
-}
-
-fn extract_bridge_info(builder: &mut BridgePropertiesBuilder, info: &LinkInfo) -> bool {
-    let mut is_bridge = false;
-    if let LinkInfo::Data(InfoData::Bridge(datas)) = info {
-        is_bridge = true;
-        for data in datas {
-            match data {
-                InfoBridge::VlanFiltering(f) => {
-                    builder.vlan_filtering(*f);
-                }
-                InfoBridge::VlanProtocol(p) => {
-                    builder.vlan_protocol(EthType::from(*p));
-                }
-                _ => {}
-            }
-        }
-    }
-    is_bridge
-}
-
-impl TryFrom<LinkMessage> for Interface {
-    type Error = InterfaceBuilderError;
-
-    fn try_from(message: LinkMessage) -> Result<Self, Self::Error> {
-        let mut builder = InterfaceBuilder::default();
-        builder.index(message.header.index.into());
-        let mut vtep_builder = VtepPropertiesBuilder::default();
-        let mut vrf_builder = VrfPropertiesBuilder::default();
-        let mut bridge_builder = BridgePropertiesBuilder::default();
-        let mut is_bridge = false;
-        builder.admin_state(if message.header.flags.contains(LinkFlags::Up) {
-            AdminState::Up
-        } else {
-            AdminState::Down
-        });
-        builder.controller(None);
-        builder.mac(None);
-
-        for attr in &message.attributes {
-            match attr {
-                LinkAttribute::Address(addr) => {
-                    builder.mac(SourceMac::try_from(addr).ok());
-                }
-                LinkAttribute::LinkInfo(infos) => {
-                    for info in infos {
-                        extract_vrf_data(&mut vrf_builder, info);
-                        extract_vxlan_info(&mut vtep_builder, info);
-                        is_bridge |= extract_bridge_info(&mut bridge_builder, info);
-                    }
-                }
-                LinkAttribute::IfName(name) => match InterfaceName::try_from(name.clone()) {
-                    Ok(name) => {
-                        builder.name(name);
-                    }
-                    Err(illegal_name) => {
-                        error!("{illegal_name:?}");
-                    }
-                },
-                LinkAttribute::Controller(c) => {
-                    builder.controller(Some(InterfaceIndex::new(*c)));
-                }
-                LinkAttribute::OperState(state) => match state {
-                    State::Up => {
-                        builder.operational_state(OperationalState::Up);
-                    }
-                    State::Unknown => {
-                        builder.operational_state(OperationalState::Unknown);
-                    }
-                    State::Down => {
-                        builder.operational_state(OperationalState::Down);
-                    }
-                    _ => {
-                        builder.operational_state(OperationalState::Complex);
-                    }
-                },
-                _ => {}
-            }
-        }
-
-        match (vrf_builder.build(), vtep_builder.build()) {
-            (Ok(vrf), Err(_)) => {
-                builder.properties(InterfaceProperties::Vrf(vrf));
-            }
-            (Err(_), Ok(vtep)) => {
-                builder.properties(InterfaceProperties::Vtep(vtep));
-            }
-            (Err(_), Err(_)) => {
-                if is_bridge {
-                    match bridge_builder.build() {
-                        Ok(bridge) => {
-                            builder.properties(InterfaceProperties::Bridge(bridge));
-                        }
-                        Err(err) => {
-                            error!("{err:?}");
-                        }
-                    }
-                }
-            }
-            (Ok(vrf), Ok(vtep)) => {
-                error!("multiple link types satisfied at once: {vrf:?}, {vtep:?}");
-            }
-        }
-        builder.build()
-    }
 }
 
 #[cfg(any(test, feature = "arbitrary"))]
